move player tabs
