+++ components/Player/Player.jsx
@@ -15,6 +15,7 @@ import {
   RecordsPage,
   CountsPage,
 } from './Pages';
+import PlayerTabs from './PlayerTabs';
 
 const getPlayerSubroute = (info, playerId) => {
   switch (info) {
@@ -35,7 +36,7 @@ const getPlayerSubroute = (info, playerId) => {
   }
 };
 
-const Player = ({ playerId, info }) => {
+const Player = ({ playerId, info, params, location }) => {
   if (!playerId) {
     return <Error />;
   }
@@ -44,6 +45,7 @@ const Player = ({ playerId, info }) => {
     <div>
       <div className={styles.header}>
         <PlayerHeader playerId={playerId} />
+        <PlayerTabs params={params} location={location} />
       </div>
       {getPlayerSubroute(info, playerId)}
     </div>
